rust, musl, dylib, wooo
This does a naughty thing in that we patch the definition of the
`x86_64-*-linux-musl` family of targets do be dynamically linked rather
than statically linked. This needs to be called out in this repos readme
and made very public so bug reports don't funnel the wrong direction.
They need to go to this repo, not the official rust repo.

Additionally added llvm-readelf and removed sccache and ccache