phase0 moved into shared build context dir
The rust bootstrap fails due to the llvm install, but its configuration
options for the initial build *is* correct. The patch can be reworked so
its less invasive. I can branch on `phiban` when there are places I need
to skip the check instead of just ripping the code out.