Python has most optional modules available
_uuid is waiting on util-linux (which provides the uuid/uuid.h import)

_tkinter should work fine, but it is heavy into X11 and years away from
    wayland support. I wasn't planning on pulling in Xorg so soon...

But other than those call outs, the rest of the deps are found
appropriately!
```
The necessary bits to build these optional modules were not found:
_tkinter                  _uuid
To find the necessary bits, look in configure.ac and config.log.

Checked 112 modules (34 built-in, 75 shared, 1 n/a on linux-x86_64, 0 disabled, 2 missing, 0 failed on import)
```